requirements.txt
python-dateutil>=1.5
PyYAML>=3.12
requests>=2.11.1
uWSGI>=2.0.15
PyMySQL>=0.7.11
Flask>=0.11.1
Flask-Script>=2.0.6
configparser>=3.5.0
behave>=1.2.5
bs4
